src/requirements.txt
pysoundfile
pydub==0.24.1
git+https://github.com/Microsoft/Cognitive-LUIS-Python.git
requests==2.22.0
numpy
aiohttp
pyaudio